scripts/markdown-to-json/pyproject.toml
[project]
name = "markdown-to-json"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "click>=8.3.0",
    "markdown>=3.9",
    "attrs>=23.0.0",
    "cattrs>=23.0.0",
    "loguru>=0.7.3",
    "python-frontmatter>=1.0.0",
    "pytest>=8.4.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["markdown_to_json"]

[project.scripts]
markdown-parser = "markdown_to_json.bin.parse_content:main"

[tool.pytest.ini_options]

[dependency-groups]
dev = [
    "ruff>=0.14.2",
]
